Feat: shooting (#13)
* Update Enemy.prefab

Added FirePoint for shooting for defaultEnemy

* ShootingMechanic

Adjusting FirePoint position, changing in-game Unity data (enemy shooting speed, bullet prefab for enemy), changing script

* DemonEnemy

Added new Demon enemy, which haunts player and deals damage upon collision

* refactor: use one script for exxploding enemy

fix: enemy fire point missing

* fix: enemy movement speed

* chore: speed is float

* fix: navmesh

* feat: add weapons

---------

+++ Assets/Scripts/Characters/CharacterStats.cs
@@ -10,6 +10,6 @@ namespace TPS.Characters
         public int MaxHealth = 100;
 
         [Tooltip("The movement speed of the character in units per second")]
-        public int MovementSpeed = 5;
+        public float MovementSpeed = 5;
     }
 }
